requirements.txt
# Salesforce Integration
simple-salesforce>=1.12.6
requests>=2.31.0

# AWS SDK
boto3>=1.34.0

# Azure SDK
azure-functions>=1.18.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0

# Analytics
pandas>=2.2.0
numpy>=2.0.0

# Web Dashboard
flask>=3.0.0
flask-cors>=4.0.0

# Streamlit Cloud Dashboard
streamlit>=1.30.0
plotly>=5.18.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.2

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
moto>=5.0.0
